+++ example-docs/sidebars.js
@@ -1,6 +1,10 @@
 module.exports = {
   someSidebar: {
     Docusaurus: ['d-s-l-test', 'd-s-l-test2', 'd-s-l-test3'],
-    Features: ['mdx'],
+    SidebarParent: [
+      {
+        SidebarChild: ['nested_sidebar_doc']
+      }
+    ],
   },
 };